+++ app/src/main/java/com/example/android_assignment2/adpater/PopMusicAdapter.kt
@@ -1,15 +1,18 @@
 package com.example.android_assignment2.adpater
 
+import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.example.android_assignment2.R
 
 class PopMusicAdapter : RecyclerView.Adapter<PopMusicViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PopMusicViewHolder {
         // here you inflate each cardview for the songs from your xml file
         // and then create your view holder passing that view you have inflated
-        return PopMusicViewHolder()
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.your_items_song, parent, false)
+        return PopMusicViewHolder(view)
     }
 
     override fun onBindViewHolder(holder: PopMusicViewHolder, position: Int) {
